Add Logging for Empty Traces (#968)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/968

Recently, we have had users seen empty traces when the system is idle leading to confusion as to whether it was caused by a bug in kineto formatting or not. This diff adds further logging to ensure that the user is aware that kineto finds no valid trace events and is expecting an empty trace.

Reviewed By: aaronenyeshi

Differential Revision: D60311331