Avoid marking every profile loop stop as Collection stage, use data available to mark errored stages. (#977)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/977

We already have the data collected to know if the collection was stopped due to `collectionDone` or `stopCollection`, the later is only set when CUPTI abruptly stops in events like not finding buffers.

We infact also set this in the itnernal Error Counters, so leverage that functionality within UST logging as well to denote a terminal stage within Kineto.

Reviewed By: aaronenyeshi

Differential Revision: D61226939

fbshipit-source-id: a4d5fa525d4457d44f0b959e4761b82de160152c