Extend CPU User Annotations to End of Profile (#986)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/986

If a CPU User Annotation doesn't end by the time the profile ends, the annotation is marked as a 0-length event. This can be annoying to look at because it seems like profiler just never got the annotation event when it did. Lets set the end time to the end of profiling.

Reviewed By: aaronenyeshi

Differential Revision: D62604717