Overlapping D2H transfer example
Summary:
This diff instantiates 6 CPU threads that schedule CUDA kernels and memory copies between CPU and GPU. For each thread there are 2 CUDA streams: one for CUDA kernels, one for memory copies. For every kernel call, we instantiate in advance a pair of inputs and outputs on the host and on the device.

The CUDA kernel needs the following inputs:
- A buffer input
- B buffer input
- C buffer output

The memory copies are as follows:
- A is in pinned memory and copied via a H2D
- B is in pageable memory and copied via H2D
- 50% of C output buffers are in pinned and 50% are in pageable

The D2H to H2D ratio in the current config is somwhere around 2:1

3 out of 6 threads are issuing transactions on the lowest CUDA stream priority and the rest of them are using default priority.

Reviewed By: xerothermic

Differential Revision: D62670279

fbshipit-source-id: e3bca9af984eeae4bce865c35663bfd53ef97ba0