Add 1ns buffer to Roctracer Events (#992)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/992

As reported in https://github.com/ROCm/roctracer/issues/105, there is an issue where event starts and ends can "tie". This can cause a visual issue in the traces. Lets add a tiny buffer so the events are separate. At the single nanosecond level, the timings are inaccurate anyways so it doesn't really hurt to add this buffer in the meanwhile. Remove/wrap in ifdef once it is issue is resolved

Reviewed By: aaronenyeshi

Differential Revision: D63296093

fbshipit-source-id: 09e313e55bbee65f5e6a4974dc52b3e0df4d5922