Do Not Cache PID/TID in Logging (#994)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/994

S451588 was caused by the LOG macro caching values that would then be copied to other processes via forking. In general, we should probably use fork handlers to clear out said variables, but from a hygiene point of view we should also not be changing control flow based on logging. For this reason, the pid/tid retrieval in logging should get the cached variable if it exists, but never do the caching itself.

Differential Revision: D63668265