Remove Stream from HIP Launches (#995)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/995

We would like for AMD traces to match CUDA matches as closely as possible. Right now we are adding stream to the CPU launches whereas in CUDA we do not. This will make zoomer try to process these events when they shouldn't. To make matters worse, the stream id that is embedded in the cpu event don't even match the GPU event. For this reason, lets get rid of them all together.

Reviewed By: aaronenyeshi

Differential Revision: D63798661

fbshipit-source-id: 0802846565b275565329e9ddcdbd7d882880b0aa