Fix Clear On Fork (#998)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/998

Threads can have the same PID and TID if forked after they have been cached. This can be dangerous especially if you have already ran a profiler workload and then decide to fork after and run another workload. To get around this, we add a routine to always clear the PID/TID when forking. Also, we make sure that the ConfigLoader will not wait on a thread when destructing since the forked thread will have no spawned child if the configuration has been already enabled.

Reviewed By: briancoutinho

Differential Revision: D63924780

fbshipit-source-id: 6ba70066ff02a263d85645287f895c999899f06c