Reset TLS on Profiling Entrance (#999)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/999

D63924780 broke some tests because of pthread_atfork having strange properties with subsequent calls. To fix this, lets deviate from this method and just reset the TLS whenever we enter the a profiling context. This will ensure that we will start "fresh" for all the PID/TID related content upon every profile. The drawbacks are:

1. 1 Extra Cache Miss per Profile - This is negligible because the cache miss is during the prepare stage for auto-trace and schedule for on-demand. To add to this, the cache miss penalty is very small

2. No Reset if Fork during Profile - If someone were to fork in the middle of a profile the TLS won't get reset. However, there are many other issues that could happen due to a fork midway through a profile such as undefined behavior with cupti, distorted profiling window etc. We shouldn't worry about this case as of today.

Reviewed By: aaronenyeshi, briancoutinho

Differential Revision: D64120658

fbshipit-source-id: d7ed8462f76dfe6042f4a9a97979fa5010cccd2e