Prevent Negative TIDs in Trace (#1002)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/1002

Right now we take the TIDs returned from CUPTI and convert them from int64_t to int32_t which results in negative tids. Because Perfetto uses int32_t in its protobuf files (according to https://github.com/google/perfetto/issues/886) we can't just convert to uint32_t because perfetto will just convert values back into their negative counterparts. To fix this, we convert to only positive values in int32_t by finding the absolute value of the original int32_t representation of the tid. We also make a special case for INT_MIN to map to 0 so that it is apparent how we handle overflow.

We add this sanitization to all tid outputs for hygiene purposes instead of just CUPTI output.

Reviewed By: ngimel

Differential Revision: D64490195

fbshipit-source-id: 63726331b90704668643806a5d9c308bc71de7c0