Fix the issue of incorrect External id for non-PyTorch activities (#1001)
Summary:
Sorry about the incomplete patch addressing the issue of setting the external ID.
I think only the PyTorch activities should set the external id to op.correlationId(), while the external ids for the runtimes and kernels should be set to op.linkedActivity()->correlationId().
However, there may be cases where the linkedActivity() for the runtime or kernel is nullptr (for example, when directly launching a kernel using a pybind mapping interface). In such cases, the external id should be empty, not set to op.correlationId().

Pull Request resolved: https://github.com/pytorch/kineto/pull/1001

Reviewed By: davidberard98

Differential Revision: D64697439

Pulled By: sraikund16

fbshipit-source-id: be8e8fcce44462767ee0205b57145a39c981f809