Remove unused-variable in kineto/libkineto/src/RoctracerLogger.cpp +3
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Differential Revision: D66330328

fbshipit-source-id: 16b9fd6fec99f0e4b18ac2814183df3541fd5cd0