Undo revert of the PR #1014 (#1017)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/1017

Reverts a rollback D66458621

Revert the Kineto rollback, this would have partially solved the issue since this part controls transmission of the metadata to the corresponding kernel. but record_param_comms in pytorch is the real issue and was still recording this metadata and would still make an invalid trace JSON when working with GPUs>30 (our truncation case).

Reviewed By: sraikund16

Differential Revision: D66475394

fbshipit-source-id: 5b781ccb27fa898a1a6496c72733f72fd31c822e