Fix CPU Trace Message
Summary: Several users have complained about how the current CPU trace iteration message is confusing. They tend to think that this means there is only one step iteration recorded. Lets change the wording to make it clear it is an entire span that is being recorded. Lets also make it clear this is for specifically CPU

Reviewed By: aaronenyeshi

Differential Revision: D66683879