Fix CPU Trace Message (#1018)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/1018

Several users have complained about how the current CPU trace iteration message is confusing. They tend to think that this means there is only one step iteration recorded. Lets change the wording to make it clear it is an entire span that is being recorded. Lets also make it clear this is for specifically CPU

Reviewed By: aaronenyeshi

Differential Revision: D66683879

fbshipit-source-id: 648a5adc625474050a7e918324dae96d5c2b6767