Add Metadata for Annotation (#1031)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/1031

We recently extended annotations for on-demand to the end of a profile if the annotation does not finish by then. This is useful to know where the tracing ended. However, it is hard to tell when looking to see if the annotation ended as the trace ended or if the annotation ended because of the tracing ended. We add this extra metadata to deliniate this

Reviewed By: sanrise

Differential Revision: D68864115

fbshipit-source-id: 6e5e36f3f50fba3ffe0f12478140e8f240b0188e