Move "baseTimeNanoseconds" and "displayTimeUnit" before "traceEvents" (#1054)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/1054

`baseTimeNanoseconds` is currently put at the end of the trace after `traceEvents` which makes efficiently parsing this trace metadata with `ijson` difficult. This diff moves "baseTimeNanoseconds" and "displayTimeUnit" before "traceEvents" - we intentionally leave `traceName` at the end of the trace for the time being to avoid having to make a more sprawling change to correctly handle trailing commas.

Reviewed By: sraikund16

Differential Revision: D70404528