Fix include in namespace libkineto (#1056)
Summary:
Thank you for maintaining Kineto and for all the work that goes into making these tools available! I’ve been using Kineto APIs and encountered an issue in ApproximateClock.h that could lead to unexpected compile errors.

https://github.com/pytorch/kineto/blob/08fcb941769f6afb6ad5792c29e1c0db891d07ab/libkineto/src/ApproximateClock.h#L21-L43

Here `#include <x86intrin.h>` is placed inside the namespace, which I think is generally discouraged in C/C++ because it causes the contents of the header—and any headers it indirectly includes—to be placed within that namespace. For example, since `ApproximateClock.h` indirectly includes `cstdlib` in the `libkineto` namespace, standard symbols like abs will be defined in `namespace libkineto` instead of the global namespace, resulting in errors such as:
```
/usr/include/c++/14.2.1/bits/std_abs.h:52:11: error: ‘abs’ has not been declared in ‘::’
   52 |   using ::abs;
```

To adhere to best practices and avoid these issues, I have moved the include directives outside of the namespace.

Pull Request resolved: https://github.com/pytorch/kineto/pull/1056

Reviewed By: aaronenyeshi

Differential Revision: D71082945

Pulled By: sraikund16

fbshipit-source-id: 85ad341e4bf18a603d8b2d9217c26f8eaccebf41