Memory Snapshot On Demand (#1067)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/1067

Kineto side of Memory Snapshot On Demand.

1. Add option to Config.cpp to enable memory snapshot
2. Add client interface API to call memory snapshot in profiler instead of typical pytorch profiler
3. Guard (2) via ifdef because we don't have profiler merged yet

NOTE: The actual logic for running the snapshot is really simple. Start->wait->export->stop. This is done intentionally. It is better to iterate and add more features as requested than to bloat a new feature with things like iterative stepping etc.

Reviewed By: sanrise

Differential Revision: D72087641