Don't Allow Profiling Thread to pick up wrong config
Summary:
Now that we have two types of on-demand threads accepting requests (GPU Traces + Memory Snapshots) we need to lock out one thread from using another's request. Right now there is a possibility that a user does memory profiling then gpu trace profiling which will cause the second trace to be invalid.

NOTE: To get both types of requests working in a single run, we will need a follow up diff in which both threads can be instantiated and tracked. Currently, if we make one type of thread, we don't allow the controller to make another. This diff is an incremental step towards getting both to work.

Reviewed By: fenypatel99

Differential Revision: D72980123

fbshipit-source-id: b4b745935495ac29e8fdf9691a5ca5e8f890a935