Move Time Converter to Configure by default (#1085)
Summary:

Right now we are setting the converter after getting timestamps for the gpuOnly case for only AMD. However, when people want to use NV only this causes an issue. Instead of adding to both codepaths lets just add a default converter. In the case that the profiler gets triggered it will overwrite the converter anyways.  This way we are always in sync and always have a converter.

Differential Revision: D74673487