Move Time Converter to Configure by default (#1085)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/1085

Right now we are setting the converter after getting timestamps for the gpuOnly case for only AMD. However, when people want to use NV only this causes an issue. Instead of adding to both codepaths lets just add a default converter. In the case that the profiler gets triggered it will overwrite the converter anyways.  This way we are always in sync and always have a converter.

Reviewed By: sanrise

Differential Revision: D74673487

fbshipit-source-id: 5e80634a491cd6ae12ea873b29d25be34293a288