Roctracer: collect cpu timestamps in approximateTime to avoid ts inversions (#1082)
Summary:
Roctracer reports gpu timestamp values is CLOCK_MONOTONIC.  This is an accurate and monotonic clock suitable for profiling.  This causes issue when trying to interface with the mixture of inaccurate and non-monotonic clocks used in kineto and pytorch.

Currently there are timestamp inversions between hip calls and autograd ranges.  This is causing headaches with analysis (as callers are exiting before their callees). The only reliable solution to avoid this is to collect cpu timestamps for hip calls in the same clock as pytorch/autograd, i.e. approximateTime().

GPU timestamps (from the accurate clock) will still need to be converted to the compromised clock.  Ideally we could do this incrementally as the kernel completions come in... in an effort to control the clock divergence.  Sadly the time converter is not available at collection time so it is not possible to compute an offset at collection time.  So all gpu timestamps will use the same time offset (post-collection) and perhaps drift terribly over long traces.

In this PR:
Collect hip call timestamps in approximateClock
Convert gpu timestamps similarly to before

Pull Request resolved: https://github.com/pytorch/kineto/pull/1082

Reviewed By: aaronenyeshi

Differential Revision: D74508159

Pulled By: sraikund16

fbshipit-source-id: a904768eec9aed98e5854faafb9be03e22eb650d