Fix fmt::format() in logging (#1098)
Summary:

Change << fmt::format() to fmt::print for *most* logging cases. We can't do it for the LOG(...) cases because it can return (void) 0 if the severity is not high enough. Luckly for us, these cases are rare as they only occur on unexpected errors. The big instances of this speedup occur in Logger.cpp and output_json.cpp and we are able to fix all instances of it.

Reviewed By: sanrise

Differential Revision: D75828524