Add toggleCollectionDynamic for XPUPTI (#1088)
Summary:
This PR adds support for toggleCollectionDynamic for Xpupti.

Since Xpupti enables/disables activities based on the session rather than the profiler instance, I cast the `IActivityProfilerSession` to an `XpuptiActivitySession`. **This assumes that all sessions are XpuptiActivitySessions**.  In upstream Pytorch, the design rule for now is that there is no simultaneous support for XPU and any other device (CUDA/Roc), so this assumption should be safe?

Pull Request resolved: https://github.com/pytorch/kineto/pull/1088

Reviewed By: aaronenyeshi

Differential Revision: D76075509

Pulled By: sraikund16

fbshipit-source-id: 3057a416fc093813a407cc6a161d8161854336cd