Update README (#1118)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/1118

### Summary

#### Overview

This diff updates the README files for Kineto and Libkineto to reflect changes and enhancements in the projects.

#### Changes to Libkineto README

*   Add information about the library's ability to output traces in various formats, including JSON for visualization in Perfetto/Chrome Trace Viewer.
*   Update the list of supported GPU architectures to include XPU and HPU, in addition to NVIDIA and AMD GPUs.

#### Changes to Kineto README

The README file for Kineto has been updated to:

*   Clarify that Kineto is a library used in the PyTorch Profiler, rather than being the profiler itself.
*   Emphasize Libkineto as the central component of Kineto for low-overhead GPU timeline tracing.
*   Remove redundant information about the goals and design principles of Libkineto.

#### New Image File

A new binary image file, `kineto_interconnect.png`, has been added to the documentation directories in both `fbcode` and `xplat` repositories. This image will likely be used to illustrate the interconnectivity of components within the Kineto project.

Differential Revision: D79186723