Introduce AIU profiler plugin, following XPU kineto plugin design (#961)
Summary:
In this PR, the AIU profiler is introduced by following kineto plugin design under libkineto/src/plugin/xpupti. The AIU profiler is build on the foundation of IBM AIU Profiling Toolkit. The plugin is disabled by default and is only activated if the system that kineto is build has the AIUPti installed and LIBKINETO_NOAIUPTI is not set to OFF.

Signed-off-by: Marcelo Amaral <marcelo.amaral@ibm.com>