Default to time based config for ondemand profiling in training when iterations don't exist (#1149)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/1149

In ActivityProfilerController.cpp scheduleTrace is only called through ondemand tracing flow. prepare_trace is what the API (kineto_shim) calls. So in the ondemand tracing path, if we see both time based and iteration based tracing in the config, fallback to time based; since in ondemand tracing that is more reliable.

The context here is that in ondemand flow, which relies on no code-instrumention via the torch.profiler context manager in python, the step function is likely not getting called and in those cases we don't want to ignore an ondemand config that has active duration ms set in the config.

Reviewed By: sraikund16

Differential Revision: D82772039

fbshipit-source-id: c38b13f4bacc060bc73fe33900b2fd25e51df450