Do not cause config getters to create IPC clients
Summary:
We may be too flexible when it comes to init'ing the IPCConfigClient if one is not found. As we have observed in prod it can cause some unknown behaviors.

So let's step back and on first principals make the flow more strict.

We always expect DaemonConfigLoader.registerFactory() to be called before any config loader thread starts, this is by design and should happen in the main thread. There on, every config_loader update thread should attempt at reading the config (if available) if the configClient exists, if it doesn't exist there is already good error handling and logging that we should leverage.

The worst thing that can happen here is one iteration of the background thread for reading config will miss a profiling config and log the error, that's far more better than debugging race conditions

Differential Revision: D84078972